main: fix quit patterns for client/server at test
Each test instantiates a new freezer server, and is
responsible for shutting that app down when the
test has finished.

The Close() method handled by the server does not
indicate that the server should close, but that
that client is Close-ing (eg. client quitting).

The S3 freezer server logic doesn't do much on Close,
if anything, so this call is simply read and logged.

Signed-off-by: meows <b5c6@protonmail.com>